% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{codebook}
\alias{codebook}
\title{V-Dem codebook}
\format{
A data frame with 828 observations and 14 variables.
}
\source{
https://www.v-dem.net
}
\usage{
codebook
}
\description{
Loads as data frame selected information from the V-Dem dataset codebook
}
\details{
For the full pdf version of the codebook and more info see https://www.v-dem.net
}
\references{
Coppedge, Michael, John Gerring, Carl Henrik Knutsen, Staffan I. Lindberg, Jan Teorell, David Altman, Fabio Angiolillo, Michael Bernhard, Agnes Cornell, M. Steven Fish, Linnea Fox, Lisa Gastaldi, Haakon Gjerløw, Adam Glynn, Ana Good God, Sandra Grahn, Allen Hicken, Katrin Kinzelbach, Kyle L. Marquardt, Kelly McMann, Valeriya Mechkova, Anja Neundorf, Pamela Paxton, Daniel Pemstein, Johannes von Römer, Brigitte Seim, Rachel Sigman, Svend-Erik Skaaning, Jeffrey Staton, Aksel Sundström, Marcus Tannenberg, Eitan Tzelgov, Yi-ting Wang, Felix Wiebrecht, Tore Wig, and Daniel Ziblatt. 2025. "V-Dem Codebook v15" Varieties of Democracy (V-Dem) Project. https://v-dem.net/data/reference-documents/
}
\keyword{datasets}
