% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{episodes}
\alias{episodes}
\title{Episodes of Regime Transformation (ERT)}
\format{
A country-year dataset (class "tibble::tbl_df") with corresponding
  to the full V-Dem data set of cases, with three dozen columns related to
  the ERT coding. Key columns include:
  
\describe{
    \item{country_id}{V-Dem numeric country identifier}
    \item{year}{Year}
    \item{dem_ep}{Binary indicator for a democratization episode}
    \item{aut_ep}{Binary indicator for an autocratization episode}
}

For information on the remaining columns, see the ERT codebook, which is 
installed along with the package. Try the following code to open it, might
not work on Windows: 

\code{browseURL(system.file("ERT_codebook.pdf", package = "ERT"))}
}
\usage{
episodes
}
\description{
Country-year data on episodes of regime transformation.
}
\details{
These data can be re-generated by running [get_eps()] with the 
default settings.
}
\examples{
data(episodes)
dplyr::glimpse(episodes)
}
\keyword{datasets}
