% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_overlap.R
\name{find_overlap}
\alias{find_overlap}
\title{Find the overlap between episodes of democratization and autocratization.}
\usage{
find_overlap(
  start_incl = 0.01,
  cum_incl = 0.1,
  year_turn = 0.03,
  cum_turn = 0.1,
  tolerance = 5
)
}
\arguments{
\item{episodes}{The outcome of get_eps(), Episodes of Regime Transformation (ERT), 
to be used for finding potential overlaps (depending on individual parameter setting). By default with standard parameters.}
}
\value{
A summary message and data frame showing the overlaps between democratization and autocratization episodes.
}
\description{
Find the overlap between episodes of democratization and autocratization.
}
\examples{
#Don't run
#Find the overlap between democratization and autocratization episodes

#overlap <- find_overlap(episodes)

}
