% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_eps.R
\name{get_eps}
\alias{get_eps}
\title{Get episodes of regime transformation (ERT)}
\usage{
get_eps(
  data = ERT::vdem,
  start_incl = 0.01,
  cum_incl = 0.1,
  year_turn = 0.03,
  cum_turn = 0.1,
  tolerance = 5
)
}
\arguments{
\item{data}{The data based on which the episodes are identified.
By default the most recent vdem data set.}

\item{start_incl}{What is the minimum annual change on V-Dem's Electoral Democracy Index (EDI) necessary to trigger an episode? 
This is the absolute value of the first difference in the EDI required for the onset 
of either a democratization (+) or autocratization episode (–).}

\item{cum_incl}{What is the minimum amount of total change on the EDI necessary to constitute a manifest episode?
A potential episode might be a period involving any amount of changes over a period following an annual change equal 
to the start inclusion (e.g. 0.01). To identify substantial changes, we set a cumulative inclusion threshold. 
This is the absolute value of the total amount of change needed on the EDI to be considered manifest.}

\item{year_turn}{What is the amount of annual change in the opposite direction to trigger the termination of an episode? 
An episode may end when the case suddenly moves in the opposite direction. 
For example, during an episode of democratization, a country may experience a sudden drop on the EDI. 
This could signal the onset of an autocratization episode. To avoid overlap between episodes, 
we set the absolute value of a change in the opposite direction on the EDI 
as a trigger for the termination of an episode. \emph{Note: Advanced users who wish to remove this criteria altogether 
should set the value of year\_turn equal to cum\_turn. 
Setting this to zero would allow for an episode to terminate when any year of no change is encountered.}}

\item{cum_turn}{What is the amount of gradual change in the opposite direction to trigger the termination of an episode?
An episode may end when the case begins moving in the opposite direction gradually. For example, 
during an episode of democratization, a country may experience a gradual drop on the EDI over a number of years 
that signals democratization has ended. This could also signal the onset of an autocratization episode. 
To avoid overlap between episodes, we set the absolute value of a gradual change in the opposite direction 
on the EDI over the tolerance period (e.g. 5 years) as a trigger for the termination of an episode.}

\item{tolerance}{What is the number of years considered as tolerance for stasis or a gradual movement in the opposite direction?
The tolerance defines the number of years an episode is allowed to remain in stasis 
(i.e. no more movements equal to the start inclusion) and/or move in the opposite direction before it is terminated. 
This parameter also defines the number of years necessary for a case to be considered a democratic breakdown or 
stabilized electoral autocracy. \emph{Therefore, care is necessary when manipulating the default value. 
This could lead to large changes in the composition of episodes. 
We set the default to 5 years because this is the typical amount for an electoral cycle for most countries.}}
}
\value{
A data frame specifying episodes of regime transformation and their outcomes in the most recent V-Dem data set.
For further details and explanations on episodes and outcomes please check the ERT codebook attached to the package 
and available here: https://github.com/vdeminstitute/ERT/blob/master/inst/ERT_codebook.pdf
}
\description{
Helps to identify episodes of democratization (liberalizing autocracy, democratic deepening) 
and autocratization (democratic regression, autocratic regression) in the most recent vdem data set. 
For further details please check the ERT codebook attached to the package 
and available here: https://github.com/vdeminstitute/ERT/blob/master/inst/ERT_codebook.pdf
}
\details{
\emph{Democratization} is an umbrella term for any movement towards demcracy - 
be it in autocracies or democracies (cf. Wilson et al., 2020).
\emph{Liberalizing autocracy} is defined as a subtype of democratiztion and specifically focuses on any movement towards democracy
which starts in autocracies. \emph{Democratic deepening} is also a subtype of democratization and
concerns all those which are already democratic and further improve their democratic traits.

\emph{Autocratization} is defined as any movement towards autocracy which starts within democracies 
or autocracies (cf. Lührmann and Lindberg, Democratization, 2019).
\emph{Democratic regression} is defined as a subtype of autocratization and specifically focuses on any movement towards autocracy
which starts in democracies. \emph{Autocratic regression} is also a subtype of autocratization and
concerns all those which are already autocratic and further decline.
}
\examples{
#Don't run
#Get the episodes with standard parameters:
#episodes <- get_eps()

}
