% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vdem}
\alias{vdem}
\title{V-Dem data}
\format{
A data frame with 27913 observations and 4607 variables.
}
\source{
https://www.v-dem.net
}
\usage{
vdem
}
\description{
Loads the most recent version of the V-Dem data set.
}
\details{
Varieties of Democracy (V-Dem) is a new approach to conceptualizing
and measuring democracy. We provide a multidimensional and disaggregated
dataset that reflects the complexity of the concept of democracy as a
system of rule that goes beyond the simple presence of elections.
The V-Dem project distinguishes between five high-level principles of
democracy: electoral, liberal, participatory, deliberative, and
egalitarian, and collects data to measure these principles.
With five Principal Investigators (PIs), thirty-nine Project Managers
(PMs) with special responsibility for issue areas,
more than thirty Regional Managers (RMs),
134 Country Coordinators (CCs), Research Assistants,
and over 4,000 Country Experts (CEs). The V-Dem project is one
of the largest social science data collection projects
focusing on research. The Headquarters is based at the
V-Dem Institute, the Department of Political Science at the
University of Gothenburg, Sweden. The most recent data set (v15)
covers 202 countries from 1789-2024 and includes 500+ indicators,
81 indices and 5 high-level indices.
For more info see https://www.v-dem.net/
}
\references{
Coppedge, Michael, John Gerring, Carl Henrik Knutsen, Staffan I. Lindberg, Jan Teorell, David Altman, Fabio Angiolillo, Michael Bernhard, Agnes Cornell, M. Steven Fish, Linnea Fox, Lisa Gastaldi, Haakon Gjerløw, Adam Glynn, Ana Good God, Sandra Grahn, Allen Hicken, Katrin Kinzelbach, Joshua Krusell, Kyle L. Marquardt, Kelly McMann, Valeriya Mechkova, Juraj Medzihorsky, Natalia Natsika, Anja Neundorf, Pamela Paxton, Daniel Pemstein, Johannes von Römer, Brigitte Seim, Rachel Sigman, Svend-Erik Skaaning, Jeffrey Staton, Aksel Sundström, Marcus Tannenberg, Eitan Tzelgov, Yi-ting Wang, Felix Wiebrecht, Tore Wig, Steven Wilson and Daniel Ziblatt. 2025. "V-Dem Country-Year Dataset v15" Varieties of Democracy (V-Dem) Project. https://doi.org/10.23696/vdemds25.
}
\keyword{datasets}
