% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ERT-package.R
\docType{package}
\name{ERT}
\alias{ERT}
\title{ERT}
\description{
A package to load, expore and work with the Episodes of Regime Transformation dataset.
}
\references{
\url{https://github.com/vdeminstitute/ERT} and \url{https://www.v-dem.net}
}
\author{
Seraphine F. Maerz (package creater and maintainer, maerz@soz.uni-frankfurt.de), Amanda B. Edgell, Joshua Krusell, Laura Maxwell, Sebastian Hellmeier
}
